drd, s390: Make the annotate_ignore_* and the pth_barrier* tests produce the same output on s390 as on other systems (not tested yet).
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@12103

+++ drd/tests/annotate_ignore_rw.c
@@ -2,13 +2,14 @@
 
 #include <assert.h>  /* assert() */
 #include <pthread.h>
+#include <stdint.h>
 #include <stdio.h>   /* EOF */
 #include <unistd.h>  /* getopt() */
 #include "../../drd/drd.h"
 
-static int s_a;
-static int s_b;
-static int s_c;
+static int8_t s_a;
+static int8_t s_b;
+static int8_t s_c;
 
 static void* thread_func(void* arg)
 {
